LocoAgent/requirements.txt
# core agent + runtime
strands-agents>=0.1.7
bedrock-agentcore>=0.1.7

# http + env + types
requests>=2.32.3
python-dotenv>=1.0.1
pydantic>=2.9.2
